Address PR feedback
- Use fixed dependency versions
- Switch from Jest to Mocha + Chai to be closer to Truffle test setup
- Remove unused tools/deploy_token_genesis.js
- Split off Contract#_ensureStateOneOf from Contract#_ensureState
- Refactor deployment_tool into multiple files
- Add `npm run test:deployment_tool` and hook into travis
- Add README to tools directory, explaining the usage of the deployment tool lib
