+++ tools/deployment_tool/index.js
@@ -1,0 +1,13 @@
+const { Contract } = require('./contract');
+const { ContractRegistry } = require('./contract_registry');
+const {
+    IncrementingAddressGenerator,
+    IncrementingNonceAddressGenerator,
+} = require('./address_generators');
+
+module.exports = {
+    Contract,
+    ContractRegistry,
+    IncrementingAddressGenerator,
+    IncrementingNonceAddressGenerator,
+};
