+++ tools/merge_chainspec_accounts.js
@@ -10,10 +10,11 @@ const chainspec = JSON.parse(fs.readFileSync(chainspecPath));
 const accounts = JSON.parse(fs.readFileSync(accountsPath));
 
 Object.entries(accounts).forEach(([address, obj]) => {
-  if (chainspec.accounts[address]) {
-    console.warn(`Account at address ${address} already exists.`);
-  }
-  chainspec.accounts[address] = obj;
+    if (chainspec.accounts[address]) {
+        console.warn(`Account at address ${address} already exists.`);
+    }
+    chainspec.accounts[address] = obj;
 });
 
+// Pretty print as JSON
 console.log(JSON.stringify(chainspec, null, 4));
